moved RecentSearchesAndFilters to independent component
