+++ src/screens/recipe-book/SearchFilterSort/RecentSearchesAndFilters/RecentSearchesAndFilters.scss
@@ -1,0 +1,8 @@
+.recent-searches-and-filters {
+    background-color: #f8f8f8;
+}
+
+.clear-or-select-all {
+    font-size: 14px;
+    color: var(--main-cta);
+}
