+++ src/screens/recipe-book/SearchFilterSort/SearchFilterSort.scss
@@ -1,8 +1,1 @@
-.recent-searches-and-filters {
-    background-color: #f8f8f8;
-}
 
-.clear-or-select-all {
-    font-size: 14px;
-    color: var(--main-cta);
-}
