+++ src/styles.scss
@@ -37,6 +37,7 @@
 @import './screens/recipe-book/RecipeLists/RecipeList/_RecipeList';
 @import './screens/recipe-book/RecipeLists/TimeDietAllergies/_TimeDietAllergies';
 @import './screens/recipe-book/RecipeLists/TimeDietAllergies/ChangeModal/_ChangeModal';
+@import './screens/recipe-book/SearchFilterSort/RecentSearchesAndFilters/RecentSearchesAndFilters';
 @import './screens/recipe-book/SearchFilterSort/SearchFilterSort';
 @import './screens/recipe-book/SearchFilterSort/SearchBarContainer/SearchBarContainer';
 @import './screens/recipe-book/SearchFilterSort/SearchFilterList/SearchFilterList';
